add download script for google drive
I've come to the conclusion that dynamically caching files on request
time is doomed to have a poor P99 latency, and will always handle large
batches of cache-misses poorly, such as when someone loads an entire
blog. Also, using the rails cache will always be loading the entire
file into memory as a string before serving it. To avoid this, I should
just use NGINX and static files. I don't need to overcomplicate it.